refactor : 기존 static front 파일 이동
