+++ web-server/api/src/main/resources/static/old/place.js
